pipeline/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
openpyxl>=3.1
pandas>=2.0
duckdb>=0.10
tqdm>=4.66
fredapi>=0.5
python-dotenv>=1.0
